fixing stuff after moving / renaming
